Introduce a proper build system.
There's a lot here and I don't feel like really describing it in detail,
but the crux of it all is: all the non-native things are built as
external projects built with toolchains and tools built before them,
also as external project; there's a bunch of aggregate targets; and also
this now does not depend on a C++ library coming from the toolchain,
instead depending on libfreestanding, which is a partial implementation
of the stdlib that will be filled in as the need arises (and that will
probably not have _any_ ABI guarantees for a _long_ time, yay).

Next steps for this branch are to actually print to serial in addition
of the EFI text protocols, so that I can try to hack together some test
targets for live QEMU runs (by parsing the console output). I should
also do some testing for conformance in libfreestanding, but as that is
a much larger project to attempt, I will ignore that for now and just
assume I can actually read the spec when I am implementing std:: (and
writing this sentence is probably a very good way to cause my hubris to
crush me in the future).

Next next steps would be to get back to actually working on the loader,
so selecting the video mode, loading the kernel and the initrd, and
entering the kernel proper, so I can start actually working on an OS
again, instead of spending many evenings on creating an awful CMake
contraption.