Do a better job of handling tag changes than EP; add prune targets.
The way ExternalProject handles git tags is really... backwards. CMake
assumes that commits pointed to by tags can change, so it re-fetches
the tag from git every time, and ends up doing a reconfigure and a
rebuild every time. The way to fix that is to use UPDATE_DISCONNECTED,
but that means it doesn't do updates even if the tag changes.

Somehow, what I did here made it sane. Add a new step to ExternalProject
that fetches the tag (because there's also issues when you change the
tag with GIT_SHALLOW being enabled; sigh), and set up the dependencies
nicely. Because the new step uses the tag name in its command, I think,
CMake rebuilds it (and its dependees) whenever the tag changes; but
UPDATE_DISCONNECTED makes sure it doesn't happen every time.

The end result is we only rebuild toolchains when they actually change,
which is so much better than I thought possible.

Also I got a suggestion to see if I can make the source/build
directories of the toolchains go away once they get installed in
`install/`; turns out that that just works with the additional fetch
magic described above, so there's now all-toolchains-prune that cleans
up the directories that are unnecessary (which should just get
re-fetched and rebuilt when a tag changes in all cases, which is the
only situation we care about).

There's nothing quite like cmake magic that actually makes
ExternalProject behave the way it 100% should behave out of the box, eh?